# mpt: do not allocate new buffer when updating dirty node