--- a/pkg/core/dao/dao.go
+++ b/pkg/core/dao/dao.go
@@ -382,11 +382,6 @@ func (dao *Simple) PutStorageItem(id int32, key []byte, si *state.StorageItem) e
 		return buf.Err
 	}
 	v := buf.Bytes()
-	if dao.MPT != nil {
-		if err := dao.MPT.Put(stKey[1:], v); err != nil && err != mpt.ErrNotFound {
-			return err
-		}
-	}
 	return dao.Store.Put(stKey, v)
 }
 
@@ -394,11 +389,6 @@ func (dao *Simple) PutStorageItem(id int32, key []byte, si *state.StorageItem) e
 // given key from the store.
 func (dao *Simple) DeleteStorageItem(id int32, key []byte) error {
 	stKey := makeStorageItemKey(id, key)
-	if dao.MPT != nil {
-		if err := dao.MPT.Delete(stKey[1:]); err != nil && err != mpt.ErrNotFound {
-			return err
-		}
-	}
 	return dao.Store.Delete(stKey)
 }
 
@@ -700,4 +690,19 @@ func (dao *Simple) StoreAsTransaction(tx *transaction.Transaction, index uint32,
 // underlying store.
 func (dao *Simple) Persist() (int, error) {
 	return dao.Store.Persist()
+}
+
+// UpdateMPT updates MPT using storage items from the underlying memcached store.
+func (dao *Simple) UpdateMPT() error {
+	var err error
+	dao.Store.MemoryStore.SeekAll([]byte{byte(storage.STStorage)}, func(k, v []byte) {
+		if err != nil {
+			return
+		} else if v != nil {
+			err = dao.MPT.Put(k[1:], v)
+		} else {
+			err = dao.MPT.Delete(k[1:])
+		}
+	})
+	return err
 }