--- a/pkg/core/mpt/trie.go
+++ b/pkg/core/mpt/trie.go
@@ -418,31 +418,29 @@ func (t *Trie) updateRefCount(h util.Uint256) int32 {
 func (t *Trie) addRef(h util.Uint256, bs []byte) {
 	node := t.refcount[h]
 	if node == nil {
+		data := make([]byte, len(bs))
+		copy(data, bs)
 		t.refcount[h] = &cachedNode{
 			refcount: 1,
-			bytes:    bs,
+			bytes:    data,
 		}
 		return
 	}
 	node.refcount++
-	if node.bytes == nil {
-		node.bytes = bs
-	}
 }
 
 func (t *Trie) removeRef(h util.Uint256, bs []byte) {
 	node := t.refcount[h]
 	if node == nil {
+		data := make([]byte, len(bs))
+		copy(data, bs)
 		t.refcount[h] = &cachedNode{
 			refcount: -1,
-			bytes:    bs,
+			bytes:    data,
 		}
 		return
 	}
 	node.refcount--
-	if node.bytes == nil {
-		node.bytes = bs
-	}
 }
 
 func (t *Trie) getFromStore(h util.Uint256) (Node, error) {
@@ -462,7 +460,8 @@ func (t *Trie) getFromStore(h util.Uint256) (Node, error) {
 		data = data[:len(data)-4]
 		node := t.refcount[h]
 		if node != nil {
-			node.bytes = data
+			node.bytes = make([]byte, len(data))
+			copy(node.bytes, data)
 			node.initial = int32(r.ReadU32LE())
 		}
 	}