I had concerns about copying when refcount is enabled, but it still seems to be better under high load and roughly the same on average. I think refcount needs more care later (e.g. using `sync.Pool` or even local `Trie` pool specially for slices contained in refcount map).

According to pprof most memory is consumed in `HashNode.EncodeBinary`. This is most likely due to expansion of `Branch` or `Extension` nodes during `Trie` update. Optimization done here is the only one which gives consistent improvements in terms of both running time and memory usage.

Other things which I tested:
1. Preallocate slice for hash nodes -- no noticeable improvements.
2. Implement `Size() int` for all nodes and preallocate -- better than master, but slower than this PR irregardless of other optimizations.

I think this should be post-preview4, to test it more together with further improvements.

Some benchmark results (inmemory db) for restore on generated dump and stats reported via `runtime.ReadMemStats` at the end. The actual gain is bigger due to statistics reported for all application, not only a `Trie`.

```
# commit, refcount enabled, blocks x tx per block
# PR, no refcount, 50x10000
TotalAlloc (MiB): 21300
Mallocs: 316552557
Frees: 301643724
PauseTotalNs: 3563883
GCCPUFraction:: 0.03032
./neogo db restore --unittest --in ./chain50x10000.acc  91,80s user 1,02s system 177% cpu 52,389 total

# master, no refcount, 50x10000
TotalAlloc (MiB): 23271
Mallocs: 325117505
Frees: 311644299
PauseTotalNs: 3783016
GCCPUFraction:: 0.03183
./neogo db restore --unittest --in ./chain50x10000.acc  95,03s user 1,10s system 179% cpu 53,671 total

# PR, with refcount, 50x10000
TotalAlloc (MiB): 21837
Mallocs: 324453766
Frees: 309332521
PauseTotalNs: 3770270
GCCPUFraction:: 0.03062
./neogo db restore --unittest --in ./chain50x10000.acc  94,84s user 0,99s system 174% cpu 55,039 total

# master, with refcount, 50x10000
TotalAlloc (MiB): 23338
Mallocs: 331209938
Frees: 312070104
PauseTotalNs: 7886987
GCCPUFraction:: 0.03189
./neogo db restore --unittest --in ./chain50x10000.acc  96,85s user 0,99s system 176% cpu 55,376 total

# PR, with refcount, 10000x50
TotalAlloc (MiB): 24814
Mallocs: 371539786
Frees: 350836724
PauseTotalNs: 6344485
GCCPUFraction:: 0.03215
./neogo db restore --unittest --in ./chain10000x50.acc  107,46s user 1,36s system 178% cpu 1:00,99 total

# master, with refcount, 10000x50
TotalAlloc (MiB): 24847
Mallocs: 377009923
Frees: 356592687
PauseTotalNs: 6340990
GCCPUFraction:: 0.03230
./neogo db restore --unittest --in ./chain10000x50.acc  107,53s user 1,28s system 180% cpu 1:00,44 total
```